+++ toolkit/scaffold/project/command.py
@@ -81,7 +81,9 @@ def generate_create_project_command(
 
         click.echo(f"Project created at:\n{os.path.abspath(project_path)}")
 
-        if launch_editor or read_user_yes_no("Launch editor?", "yes"):
+        if project_path != "." and (
+            launch_editor or read_user_yes_no("Launch editor?", "yes")
+        ):
             click.edit(
                 editor=read_user_choice(
                     "Which editor to use?",