+++ toolkit/template/scaffold/project/python/package/{{project_slug.snake_case}}/config/__init__.py
@@ -2,6 +2,8 @@
 
 from pathlib import Path
 
-DEFAULT_CONFIG_FILE = Path.home() / ".config" / "{{ project_slug.kebab_case }}" / "config.yaml"
+DEFAULT_CONFIG_FILE = (
+    Path.home() / ".config" / "{{ project_slug.kebab_case }}" / "config.yaml"
+)
 
 DEFAULT_CONFIG_FILE.parent.mkdir(parents=True, exist_ok=True)