+++ toolkit/template/scaffold/project/python/package/{{project_slug.snake_case}}/config/settings.py
@@ -4,7 +4,6 @@ from typing import Any, Dict
 import yaml
 from pydantic import BaseSettings
 from pydantic.env_settings import SettingsSourceCallable
-
 from {{project_slug.snake_case}}.logger import logging
 
 log = logging.getLogger(__name__)